[不具合][アップグレード画面]  アップル審査リジェクト対応
Guideline 2.1 - Performance - App Completeness

We discovered one or more bugs in your app.

Specifically, we are unable to purchase the Standard Plan in-app purchase because the purchase button is unresponsive.

Review device details:

- Device type: iPhone
- OS version: iOS 14.4

Next Steps

Please run your app on a device to reproduce the issues, then revise and submit your app for review.

If at first you're unable to reproduce the issue, try the following:

- For new apps, uninstall all previous versions of your app from a device, then install and follow the steps to reproduce.
- For app updates, install the new version as an update to the previous version, then follow the steps to reproduce.

If we misunderstood the intended behavior of your app, please reply to this message in Resolution Center to provide information on how these features were intended to work.

Resources

- For information about testing apps and preparing them for review, see Technical Note TN2431: App Testing Guide.
- To learn about troubleshooting networking issues, see About Networking.

Please see attached screenshots for details.